Fix printing without watermark
Entering the group outside of the loop were making the app wait forever if the watermarking is disabled.
Correct behavior is to enter the group for each image to be watermarked and leave the group after each image is watermarked. Not just enter once and leave once when all is done.
Also if no watermark no need to enter the group.